- handle initial text in an org-file which isn't part of a headline
Initial text in a file that isn't a headline (e.g. preceded with a "* ") will
be treated as the body (notes) text for an empty headline (i.e. a line
containing only "* ").  In other words, the text will be parsed as if there
were an empty headline inserted before the non-headline lines.

- added test for this initial non-headline condition